fix upgrade bug

+++ app/src/main/java/com/savor/ads/utils/AppUtils.java
@@ -628,15 +628,17 @@ public class AppUtils {
                 }
 
                 // 删除sdcard中form表单处理可能遗留的临时文件
-                String sdcardPath = getExternalSDCardPath();
-                File sdcardDirectory = new File(sdcardPath);
-                if (sdcardDirectory.exists()) {
-                    File[] files = sdcardDirectory.listFiles();
+                String sdcardPath = getSDCardPath();
+                if (!TextUtils.isEmpty(sdcardPath)) {
+                    File sdcardDirectory = new File(sdcardPath);
+                    if (sdcardDirectory.exists()) {
+                        File[] files = sdcardDirectory.listFiles();
 
-                    if (files != null) {
-                        for (File file : files) {
-                            if (file.isFile() && file.getName().startsWith("MultiPart")) {
-                                com.savor.ads.utils.FileUtils.deleteFile(file);
+                        if (files != null) {
+                            for (File file : files) {
+                                if (file.isFile() && file.getName().startsWith("MultiPart")) {
+                                    com.savor.ads.utils.FileUtils.deleteFile(file);
+                                }
                             }
                         }
                     }
